Move some stuff to core-android module
