+++ client-data/src/main/java/com/jraska/github/client/users/RepoDetailViewModel.kt
@@ -7,7 +7,7 @@ import com.jraska.github.client.Urls
 import com.jraska.github.client.analytics.AnalyticsEvent
 import com.jraska.github.client.analytics.EventAnalytics
 import com.jraska.github.client.rx.AppSchedulers
-import com.jraska.github.client.rx.RxLiveData
+import com.jraska.github.client.core.android.rx.RxLiveData
 
 class RepoDetailViewModel constructor(
   private val usersRepository: UsersRepository,
