+++ client/src/main/java/com/jraska/github/client/ui/UriHandlerActivity.kt
@@ -1,7 +1,7 @@
 package com.jraska.github.client.ui
 
 import android.os.Bundle
-import com.jraska.github.client.UriHandlerViewModel
+import com.jraska.github.client.core.android.UriHandlerViewModel
 import com.jraska.github.client.inputUrl
 import com.jraska.github.client.viewModel
