+++ core-android/src/main/java/com/jraska/github/client/core/android/LinkLauncher.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client
+package com.jraska.github.client.core.android
 
 import android.app.Activity
 import okhttp3.HttpUrl
