+++ core-android/src/main/java/com/jraska/github/client/core/android/UriHandlerViewModel.kt
@@ -1,6 +1,7 @@
-package com.jraska.github.client
+package com.jraska.github.client.core.android
 
 import androidx.lifecycle.ViewModel
+import com.jraska.github.client.DeepLinkHandler
 import okhttp3.HttpUrl
 
 class UriHandlerViewModel constructor(private val deepLinkHandler: DeepLinkHandler) : ViewModel() {
