+++ core-android/src/main/java/com/jraska/github/client/core/android/rx/RxLiveData.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.rx
+package com.jraska.github.client.core.android.rx
 
 import androidx.lifecycle.LiveData
 import io.reactivex.Maybe
